adds a temporary filter object component
