+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Components/Collections/GetCollectionObjects.cs
@@ -1,0 +1,139 @@
+using System.Runtime.InteropServices;
+using Grasshopper.Kernel;
+using Speckle.Connectors.GrasshopperShared.HostApp;
+using Speckle.Connectors.GrasshopperShared.Parameters;
+using Speckle.Connectors.GrasshopperShared.Properties;
+using Speckle.Sdk;
+using Speckle.Sdk.Models;
+
+namespace Speckle.Connectors.GrasshopperShared.Components.Collections;
+
+/// <summary>
+/// Given a collection, this component will output the objects in the collection or any child collection according to input paths
+/// </summary>
+[Guid("77CAEE94-F0B9-4611-897C-71F2A22BA311")]
+public class GetCollectionObjects : GH_Component
+{
+  public override Guid ComponentGuid => GetType().GUID;
+  protected override Bitmap Icon => Resources.speckle_objects_query;
+
+  public GetCollectionObjects()
+    : base(
+      "GetCollectionObjects",
+      "gcO",
+      "Retrieves the objects inside a Speckle collection at the specified path",
+      ComponentCategories.PRIMARY_RIBBON,
+      ComponentCategories.COLLECTIONS
+    ) { }
+
+  protected override void RegisterInputParams(GH_InputParamManager pManager)
+  {
+    pManager.AddParameter(
+      new SpeckleCollectionParam(),
+      "Collection",
+      "C",
+      "Collection to retrieve objects from",
+      GH_ParamAccess.item
+    );
+
+    pManager.AddTextParameter(
+      "Collection Path",
+      "C",
+      "Get the Speckle objects in the subcollection indicated by this path",
+      GH_ParamAccess.item
+    );
+
+    Params.Input[1].Optional = true;
+  }
+
+  protected override void RegisterOutputParams(GH_OutputParamManager pManager)
+  {
+    pManager.AddParameter(
+      new SpeckleObjectParam(),
+      "Objects",
+      "O",
+      "The objects in the input collection that match the queries",
+      GH_ParamAccess.tree
+    );
+  }
+
+  protected override void SolveInstance(IGH_DataAccess dataAccess)
+  {
+    SpeckleCollectionWrapperGoo collectionWrapperGoo = new();
+    dataAccess.GetData(0, ref collectionWrapperGoo);
+
+    if (collectionWrapperGoo.Value == null)
+    {
+      return;
+    }
+
+    string path = "";
+    dataAccess.GetData(1, ref path);
+
+    // filter by collection path
+    // Note: the collection paths selector will omit the target collection from the path of nested collections.
+    // the discard ("_objects") will be used to indicate objects found directly in the target collection.
+    List<SpeckleObjectWrapper> filteredObjects = new();
+    SpeckleCollectionWrapper? targetCollectionWrapper = null;
+    if (!string.IsNullOrEmpty(path))
+    {
+      targetCollectionWrapper =
+        path == "_objects" ? collectionWrapperGoo.Value : FindCollection(collectionWrapperGoo.Value, path);
+      filteredObjects = targetCollectionWrapper
+        .Collection.elements.Where(e => e is SpeckleObjectWrapper)
+        .Select(e => (SpeckleObjectWrapper)e)
+        .ToList();
+    }
+    else
+    {
+      filteredObjects = GetAllObjectsFromCollection(collectionWrapperGoo.Value).ToList();
+    }
+
+    // Set output objects
+    if (targetCollectionWrapper?.Topology is string topology && !string.IsNullOrEmpty(topology))
+    {
+      var tree = GrasshopperHelpers.CreateDataTreeFromTopologyAndItems(topology, filteredObjects);
+      dataAccess.SetDataTree(0, tree);
+    }
+    else
+    {
+      dataAccess.SetDataList(0, filteredObjects);
+    }
+  }
+
+  private IEnumerable<SpeckleObjectWrapper> GetAllObjectsFromCollection(SpeckleCollectionWrapper collectionWrapper)
+  {
+    foreach (Base element in collectionWrapper.Collection.elements)
+    {
+      switch (element)
+      {
+        case SpeckleCollectionWrapper childCollectionWrapper:
+          GetAllObjectsFromCollection(childCollectionWrapper);
+          break;
+        case SpeckleObjectWrapper objectWrapper:
+          yield return objectWrapper;
+          break;
+      }
+    }
+  }
+
+  private SpeckleCollectionWrapper FindCollection(SpeckleCollectionWrapper root, string unifiedPath)
+  {
+    // POC: SpeckleCollections now have a full list<string> path prop on them always. Is this easier to use?
+    List<string> paths = unifiedPath.Split([Constants.LAYER_PATH_DELIMITER], StringSplitOptions.None).ToList();
+    SpeckleCollectionWrapper currentCollectionWrapper = root;
+    while (paths.Count != 0)
+    {
+      currentCollectionWrapper = currentCollectionWrapper
+        .Collection.elements.OfType<SpeckleCollectionWrapper>()
+        .First(col => col.Collection.name == paths.First());
+      paths.RemoveAt(0);
+      if (paths.Count == 0)
+      {
+        return currentCollectionWrapper;
+      }
+    }
+
+    throw new SpeckleException("Did not find collection");
+  }
+}
