+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Properties/Resources.resx
@@ -175,10 +175,4 @@
   <data name="speckle_properties_query" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\speckle_properties_query.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
-  <data name="speckle_query_equals" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\speckle_query_equals.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="speckle_query_wildcard" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\speckle_query_wildcard.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
 </root>
