+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Speckle.Connectors.GrasshopperShared.projitems
@@ -11,11 +11,12 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Components\BaseComponents\SpeckleScopedTaskCapableComponent.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\BaseComponents\SpeckleTaskCapableComponent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\BaseComponents\Operator.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\BaseComponents\ValueSet.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Collections\CollectionPathsSelector.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Collections\CreateCollection.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Collections\ExpandCollection.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Components\Collections\FilterObjectsByCollectionPaths.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Objects\QuerySpeckleObjects.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\ComponentUtils.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\GhContextMenuButton.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\GhContextMenuButtonAttributes.cs" />
@@ -42,7 +43,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\Helpers.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\SpeckleResource.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\SpeckleResourceBuilder.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Parameters\SpeckleQueryParam.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Parameters\SpeckleMaterialWrapper.ModelObjects.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Parameters\SpeckleMaterialWrapper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Properties\Resources.Designer.cs" />
